Fix contact/physbone receivers on nonlocal avatars
This fixes two bugs resulting in contacts and physbone parameters not
working on nonlocal avatars:

1. Because the `emulator` field is null on nonlocal avatars,
   `assignPhysBoneParameters()` was not being called.  Here, we update
   the condition to consult the parent avatar for nonlocal clones.

2. The parameter assignment seems to happen after the initial value of
   the contact is set by the VRCSDK, so the accessor was not being
   consulted until the value of the contact changed at least once. Here,
   we assign the parameter value explicitly when we change out the
   accessor to patch over this race condition. There's probably a
   cleaner solution, but I don't understand the order of events well
   enough to propose it.

Note: I've not fully tested the PhysBone path here, but it shouldn't
break... I think?

Note[2]: Changing the displayed layer on the nonlocal clone will also
reset the parameter values. I'm not sure how to avoid this.