Mitigate logspam from parameter drivers
When in play mode without the emulator active (e.g. when uploading),
parameter drivers trigger a lot of error-level logspam. This is a problem,
because it will trigger error pause (and confuse anyone who doesn't know
that error pause is a thing).

This change reduces the loglevel to warning, and avoids issuing the warning
more than once per avatar per play mode transition.