Invoke avatar preprocessing hooks when initializing the emulator
When writing editor scripts for avatars that run via build hooks, it can
be helpful to have them work when the Av3Emulator is running as well.
However, currently the Av3Runtime reads a bunch of data from the avatar
descriptor in Awake, and there's not a reliable way for external tools
to run prior to this point (without modifying the scene permanently).

This change simply invokes the VRCSDK's OnPreProcessAvatar hooks
automatically when Av3Runtime initializes to resolve this issue.